requirements.txt
Django==5.1.3
Django==3.2.25
djangorestframework==3.12.4
mysqlclient==2.2.6

